Revert "Install i3 stuff alongside gnome"
This reverts commit a7d1dd6a3263671507b384dd377410d86d35f0d3.

I haven't actually used i3 for a very long time. I added it to the
ansible as a courtesy to myself but I don't think I've used it for a
combined total of more than a couple minutes. Just not my thing anymore;
I've come to like a desktop environment with features.