Add support for RHEL Workstation
I actually went and bought RHEL7 Workstation for my work machine in an
attempt to make it as stable and long-term as possible.

Most of the playbook could remain the same, as the Fedora support
already makes things pretty close, but some fringe things like Lollypop
needed to change since there isn't an EL7 package for it.

w.r.t. lollypop: there is a flatpak, and RHEL7W does come with Flatpak,
so I went with that. It's a bit clunky to work with but now that it's in
Ansible I hardly care about usability tbh.

Other things were just too ancient to be useful, like PostgreSQL and
Erlang (Erlang was at R16!!!!). For these I just pulled them in from
third-party, official repos.

Also huge s/o to Sublime for having an EL7 package in the main repo. <3