Stop hacking apart GNOME
These removals were causing so many log errors and failures, as turns
out GNOME expects most other GNOME things to exist.

GNOME Software, in particular, was the cause of a fair bit of strife.
And to even launch the contacts app we need the evolution libraries and
units, so... Best to just leave it alone. GNOME is GNOME and best to not
try and make it something it's not.