Add a quick playbook for installing libva
I can never remember which package to install to make this work. Since I
don't want my standard playbooks to be hardware-specific, I'm putting
this in its own little optional playbook, just so it's documented but
still separate and opt-in, so it can be ignored for headless systems or
other weirder hardware.