Remove Gnome podcasts and flatpak remote
The flatpak stuff isn't included until Ansible 2.6, and Ubuntu 18.04
only includes 2.5, which means the playbooks won't run *at all*, even
with a conditional, as the task module is not found.

Since I never really got into Gnome Podcasts (Pocketcasts is still just
too good), and it's the only thing installed with flatpak, we'll just
remove the flatpak-specific tasks and Podcasts.