Remove support for RHEL and CentOS
I used RHEL Workstation briefly on my work laptop but was ultimately
frustrated at how old everything was, so gave up and ended up back on
Fedora. Maintaining support for RHEL has proved annoying, so I'm just
going to remove it.

This also simplifies a bunch of stuff that we had to "side-load" like
Erlang, since the in-repo stuff is so old.