Consolidate OS-specific playbooks and roles
This combines all playbooks and roles into generic ones, instead of
having separate roles for separate operating systems, and just ensures
`when` clauses are present for all tasks that are specific to a
particular platform.

Additionally, we drop our clumsy `is_*` variables and just use the
Ansible fact values. I'm not sure why we didn't just do this to begin
with.