Switch to upstream Docker CE repo on Fedora
I tried to build a multistage dockerfile just now and it failed and I
realised Fedora ships a very old Docker version (1.13), which predates
the shift to date-based versions. Multistage builds were introduced in
17.x, meaning Fedora's docker is at least 2 years out of date.

Also Docker are equally shit, as expected, and have a Fedora 30 repo but
no packages in it except containerd. So whatever. I give up. Software is
impossible to use.