Only use the VSCode snap
Microsoft did a sin by shipping a repo file with VSCode in the same RPM
as the editor, so it's literally impossible to cleanly automate the repo
installation because then you end up with two, or a constant fight of
two things overwriting each other's repo config changes.

This just consolidates and installs VScode exclusively in the snap. It
works fine.