Provisional RHEL 8 support
This mostly works to get our roles genericized enough to work with
RedHat Enterprise Linux Developer Workstation 8. Some packages are
simply missing (though EPEL and RPMFusion really help there), so this is
"provisional" support due to that.

A lot of this work is only possible because of RHEL 8. RHEL 7 was
branched off a version of Fedora that came out before these playbooks
were even really a thing, so it was always difficult to support both.
Now Fedora and RHEL are in that "close enough" state where we can target
tasks to both pretty easily.

Currently the packages that are still missing are:

- ShellCheck
- ag (the_silver_searcher)
- lollypop
- docker

We have replaced docker with podman for now, as podman gets us 80%
there and will be mostly fine for now. It's also "possible" to install
Docker CE for CentOS on RHEL if you make dnf a bit unhappy and pass some
forceful flags when installing, so hopefully it won't come to that but
it's an option...

Lollypop is one of those apps that I have stopped loving as development
has continued to add piles of unnecessary features while stripping the
UI of any sane navigation hierarchy. While it's unfortunate RHEL cannot
install it currently, it's perhaps the kick in the ass I need to find a
decent alternative.

ag and shellcheck are unfortunate omissions. For now I will compile them
from source manually, but hopefully they show up in EPEL or otherwise
soon.