Remove dotnet sdk
I enjoyed a lot of things about C# and .NET, but ultimately true
cross-platform development is just not there yet. Many things are very
difficult or impossible to do without a Visual Studio installation
still, and there are still a lot of clunky or weird Windows-isms that
make a Linux-based development environment not particularly fun.

Maybe one day, but I don't think today.