Automate asdf plugin installation
This moves asdf into its own file for cleanliness and automates the
installation of all the plugins we commonly need.

This does _not_ install any versions by default, intentionally. Doing so
gets out of date incredibly quickly, so installing versions of things
manually is still reasonable I think.

But the plugins themselves are evergreen and annoying to remember.