OpenSuse
I've been use SLES and OpenSuse a lot more recently and... it's pretty
good. It took a bit to get used to it, just because it is kind of a
weird hybrid of old-school linux stuff and modern packages.

I've taken the plunge on my linux laptop now though, and am now an
OpenSuse Tumbleweed user.

To simplify these playbooks and avoid maintaining a fourth target
operating system (lol), I just deleted everything else that wasn't
OpenSuse-compatible or macOS, which now matches the OS's I actually use
day-to-day.