Replace Viscocity/OpenVPN with Tailscale
I've adopted Tailscale in place of my old OpenVPN networks and it's
life-changing how good it is.

I wrote all the Linux parts of this off the top of my head, so there's
like a 50/50 chance it's correct or works. There is no Homebrew/macOS
component to this unfortunately as it seems right now to only be
installed via the App Store...