Remove Ubuntu support
Realistically, 95% of my computer use these days is macOS. I still have
a laptop with Fedora on it that I use rarely. I haven't had any
development machines with Ubuntu on them for a while now. Maintaining
all of these various tasks for three OS's when I only use one of them
regularly is a pain. Let's just simplify all of this.