Support Debian Stable on a best-effort basis
Ubuntu and Debian support is easy to double-up on, since the majority of
Ubuntu's package set is just upstream Debian stuff, or at least is named
the same.

Note: at this point only the headless roles are of concern, as raw
Debian on the desktop hasn't really been a thing I've done in a long
time. Maybe one day, but not today.

The only thing that didn't "just work" was `perf`, which Ubuntu chooses
to ship in some weird metapackage instead of using the upstream Debian
package for it. This is the only Ubuntu-specific override we really
needed.

Neovim in Debian Stable was also too old to work with most of my config,
so I added separate neovim tasks to compile it from source on Linuxes.
Seems like something I might want on all OS's, but for now we're keeping
the Homebrew version out of convenience (and the fact it's generally
always up to date).