Fix neovim installation tasks
Neovim's makefile seems fucked because `make install` appears to
recompile everything. This meant not only that a bunch of the generated
files became owned as `root` in the build dir (as we were elevating to
install to a global prefix), but also it wouldn't recompile with our
release target flag the second time, so we'd end up with a debug build.

A couple things to change here:

1. Just install to a local prefix. We just recently started putting
   stuff in `~/.local/bin` so why not just use that.
2. Just compile once, with all the flags. Since `make install` seems to
   just compile anyway, we can just run that instead of two discrete
   steps and everything works and we get our proper release build.