Add a very basic Windows setup script
This is the very beginnings of a Windows 11 setup script. These handful
of basic settings are the bare minimum, and there's a tonne more that
needs to be done, but I just haven't got around to finding the right
registry keys to set yet.

This at least gets it in place so we can slowly add to it over time.