Install Docker on Fedora instead of podman
I love podman, and really hope one day it rivals the quality of the
Docker tooling, but right now there's just a bit too much
incompatibility and jank that I can't justify using it for anything
important, especially for my job.

Consider this a TODO: replace docker with podman as soon as we can
without losing functionality or compatibility with other tools.