Remove Ubuntu workstation support
It's been a long time since I've used an Ubuntu desktop computer. The
closest I've come to running Ubuntu in the last few years has been in
headless VMs, which are now implicitly covered by the Debian support.

Thus, Ubuntu is no longer an "official" target of these playbooks --
only Debian. They may also work on Ubuntu, but that is unsupported and
not a goal of the project anymore.